refactor: 리뷰 반영 - domain 하위 패키지로 이동
